Update stories store to redux
