+++ src/redux/rootReducer.ts
@@ -2,9 +2,11 @@ import { combineReducers } from "@reduxjs/toolkit";
 
 // reducers
 import userReducer from "./features/user";
+import storiesReducer from "./features/stories";
 
 const rootReducer = combineReducers({
   user: userReducer,
+  stories: storiesReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
